Fix instructions for GIT_TOKEN/App Store Connect
The existing instructions did not work for me. In particular the definition of GIT_TOKEN was incorrect. My change adds a link to the current Fastlane match documentation that spells out what GIT_TOKEN should be, and those instructions worked correctly for me. I also added an extra step that seems necessary (the Fastlane deployment failed without it) which is to create the App in the App Store Connect site ahead of time.