Docs examples: Generate unique cache keys (#200)
* Docs examples: Generate unique cache keys

If the key is not unique for each build, the cache will only get updated for the first time the job is run. By generating a unique cache key, every time the job is run, a new cache entry will be created, and as per GitHub docs: "If there are no exact matches, the action searches for partial matches of the restore keys. When the action finds a partial match, the most recent cache is restored to the path directory."

* Add missing dash at end of restore key

Not strictly necessary as far as I understand, but a good practice anyway.

Co-authored-by: David Finol <davidmfinol@gmail.com>

Co-authored-by: David Finol <davidmfinol@gmail.com>