Use version from build in steam-deploy example
As you recommend to use the semantic versioning in the documentation of the builder, it might be useful to show how to use this version in the steam-deploy example. Especially since the currently used method creates different version names.

Additionally I added the `needs` section to the deployment job as it shouldn't run in parallel with the build job.