Convert versions page to actual mdx documentation
This works fine, but the layout is a bit broken. I assume this is what we will want if we get rid of antD as the tables could fit docs website style directly instead.
I'm pushing this to its own branch so we can have a preview, but for now, I'm not making this part of `docusaurus` branch.